renamed make to parse make
